requirements.txt
numpy
PyExecJS
requests
selenium
